@romstad 2022-02-02
This looks like a good idea. Unfortunately, there are some merge conflicts, and I don't have time to dive into them now. I'll have a look later.